Return process Stderr in shellRun error
exec.Cmd.Output returns only the Stdout of the executed process, thus
the actual error message was being lost. In cases where the subprocess
return a exec.ExitError use exec.ExitError.Stderr as the ouput.

It should've been possible to use exec.Cmd.CombinedOutput that returns
both Stdout and Stderr in an unified way, but with this a successful
command that also returns something in Stderr will interpolate Stderr
in the returned string.